Remove warning: comparison between signed and unsigned integer expressions [-Wsign-compare]

+++ src/clustering.cpp
@@ -326,7 +326,7 @@ ClusterTreeBuilder::clean_recursive(ClusterTree& current) const
   algo->clean(current);
   if (!current.isLeaf())
   {
-    for (size_t i = 0; i < current.nbChild(); ++i)
+    for (int i = 0; i < current.nbChild(); ++i)
     {
       if (current.getChild(i))
         clean_recursive(*((ClusterTree*)current.getChild(i)));
